requirements.txt
matplotlib>=3.3.2
numpy>=1.19.2
pandas>=1.1.4
PyYAML>=5.3.1
